refactor(copyright): split pattern extract modules
